+++ src/TimeViewer.tsx
@@ -9,8 +9,9 @@ import { CurrentDuration, mapToCurrentDuration } from "./mapToCurrentDuration"
 import { TimerState } from "./timerReducer"
 import { useObservable } from "./useObservable"
 import { createCache } from "./util/createCache"
+import { floor } from "./util/floor"
 import { formatDuration } from "./util/formatDuration"
-import { floor, interval } from "./util/interval"
+import { interval } from "./util/interval"
 
 export function TimeViewer({
   timerState$,
